Avoid using self as key for dispatch_set_specific
I know that the documentation for `dispatch_queue_set_specific` says that keys are only compared as pointers, but using `self` as the key causes Xcode 8’s memory debugger to show `dispatch_set_specific` is strongly retaining self.  I know its not the case, but it makes it hard to track down other memory retention issues.

This commit changes the key from `self` to `_workQueue`.