Remove unused code
In d3beb35574d72ca6fcb885ce87bc6e12ee95c84e creating the connection message was
extracted into the SRHTTPConnectMessageCreate function; However, generating the
secKey was pulled *outside* of that extracted method and passed in as a
parameter.  It looks like in the extraction process the now redundant
and unused key generation code was left in the extracted function.